Added runCoverage test
