+++ test/artifacts/doc-content.md
@@ -2,13 +2,15 @@
 
 [comment]: # (test start {"id":"search-kittens"})
 
-[Doc Detective documentation](http://doc-detective.com) is split into a few key sections:
-[comment]: # (step {"action":"goTo", "url":"http://doc-detective.com"})
+[Doc Detective documentation](https://doc-detective.com) is split into a few key sections:
+
+[comment]: # (step {"action":"goTo", "url":"https://doc-detective.com"})
 
 -   The landing page discusses what Doc Detective is, what it does, and who might find it useful.
 -   [Get started](https://doc-detective.com/get-started.html) covers how to quickly get up and running with Doc Detective.
     [comment]: # (step {"action":"goTo", "url":"https://doc-detective.com/get-started.html"})
 -   The [references](https://doc-detective.com/reference/) detail the various JSON objects that Doc Detective expects for configs, test specifications, tests, actions, and more. Each object schema includes an object description, field definitions, and examples.
+
     [comment]: # (step {"action":"goTo", "url":"https://doc-detective.com/reference/"})
     [comment]: # (step {"action":"goTo", "url":"https://doc-detective.com/reference/schemas/config.html"})
     [comment]: # (step {"action":"find", "selector":"h2#description", "matchText":"Description"})
